New route for manifest editor
